refactor: editor -> frontend
