+++ src/frontend/plugins/box/index.tsx
