+++ src/frontend/plugins/box/renderer.tsx
