+++ src/frontend/plugins/common-components.tsx
