+++ src/frontend/plugins/equations/editor.tsx
