+++ src/frontend/plugins/equations/index.ts
