+++ src/frontend/plugins/equations/renderer.tsx
