+++ src/frontend/plugins/equations/sign.tsx
