+++ src/frontend/plugins/registry.ts
