+++ src/frontend/plugins/serlo-table/index.tsx
