+++ src/frontend/plugins/serlo-table/renderer.tsx
